gets contianers
